requirements.txt
Django~=5.1.3
python-dotenv~=1.0.1

Pillow~=11.0.0
djangorestframework~=3.15.2